refactor(qpl): create question importer class
